requirements.txt
PyMySQL==1.0.2
Flask==2.0.2
Flask-Cors==3.0.10
line-bot-sdk==2.0.1